chore: clean up the repository and fix multiple scans of startls services
